Use webhook message body hash as id
