+++ lib/webhookPublisher/publisher.ts
@@ -1,10 +1,9 @@
 import { log } from '@charmverse/core/log';
 import type { Space } from '@charmverse/core/prisma';
 import { prisma } from '@charmverse/core/prisma-client';
-import { v4 } from 'uuid';
 
 import { addMessageToSQS } from 'lib/aws/SQS';
-import type { WebhookEventNames, WebhookEvent, WebhookPayload } from 'lib/webhookPublisher/interfaces';
+import type { WebhookEvent, WebhookPayload } from 'lib/webhookPublisher/interfaces';
 
 const SQS_QUEUE_NAME = process.env.SQS_WEBHOOK_PUBLISHER_QUEUE_NAME;
 
@@ -43,49 +42,13 @@ async function fetchSpaceWebhookSubscriptionStatus(spaceId: Space['id'], scope: 
   return webhookSubscription;
 }
 
-// export async function publishWebhookEvent<T = WebhookEventNames>(spaceId: string, event: WebhookEvent<T>) {
-//   try {
-//     if (!SQS_QUEUE_NAME) {
-//       throw new Error('Webhook SQS env var missing');
-//     }
-//     // Find if the space is subscribed to an event name or name space
-//     const subscription = await fetchSpaceWebhookSubscriptionStatus(spaceId, event.scope);
-
-//     // If no subscription, we stop here
-//     if (!subscription || !subscription.space.webhookSubscriptionUrl || !subscription.space.webhookSigningSecret) {
-//       return;
-//     }
-
-//     const webhookPayload: WebhookPayload = {
-//       event,
-//       createdAt: new Date().toISOString(),
-//       spaceId,
-//       webhookURL: subscription.space.webhookSubscriptionUrl,
-//       signingSecret: subscription.space.webhookSigningSecret
-//     };
-
-//     // Add the message to the queue
-//     await addMessageToSQS(SQS_QUEUE_NAME, JSON.stringify(webhookPayload));
-//     log.debug(`Sent event to webhook queue: "${event.scope}"`, {
-//       spaceId,
-//       createdAt: webhookPayload.createdAt,
-//       webhookURL: webhookPayload.webhookURL
-//     });
-//   } catch (e) {
-//     log.warn('Error while publishing webhook event. Error occurred', { error: e, scope: event.scope, spaceId });
-//   }
-// }
-
 export async function publishWebhookEvent(spaceId: string, event: WebhookEvent) {
   try {
     if (!SQS_QUEUE_NAME) {
       throw new Error('Webhook SQS env var missing');
     }
 
-    const webhookMessageId = v4();
-
     const webhookPayload: WebhookPayload = {
-      id: webhookMessageId,
       event,
       createdAt: new Date().toISOString(),
       spaceId,
@@ -101,12 +64,6 @@ export async function publishWebhookEvent(spaceId: string, event: WebhookEvent) 
       webhookPayload.webhookURL = subscription.space.webhookSubscriptionUrl;
       webhookPayload.signingSecret = subscription.space.webhookSigningSecret;
     }
-
-    await prisma.webhookMessage.create({
-      data: {
-        id: webhookMessageId
-      }
-    });
 
     // Add the message to the queue
     await addMessageToSQS(SQS_QUEUE_NAME, JSON.stringify(webhookPayload));
