+++ serverless/webhook/worker.ts
@@ -1,167 +1,14 @@
+import crypto from 'crypto';
+
 import { log } from '@charmverse/core/log';
 import { prisma } from '@charmverse/core/prisma-client';
 import type { SQSBatchItemFailure, SQSBatchResponse, SQSEvent, SQSRecord } from 'aws-lambda';
 
-import { createInlineCommentNotification } from 'lib/notifications/createInlineCommentNotification';
-import { createCardNotification, createDocumentNotification } from 'lib/notifications/createNotification';
-import { extractMentions } from 'lib/prosemirror/extractMentions';
-import type { PageContent } from 'lib/prosemirror/interfaces';
+import { createNotificationsFromEvent } from 'lib/notifications/createNotificationsFromEvent';
 import { signJwt } from 'lib/webhookPublisher/authentication';
-import { whiteListedWebhookEvents, WebhookEventNames } from 'lib/webhookPublisher/interfaces';
-import type { WebhookEvent, WebhookPayload } from 'lib/webhookPublisher/interfaces';
+import type { WebhookPayload } from 'lib/webhookPublisher/interfaces';
+import { whiteListedWebhookEvents } from 'lib/webhookPublisher/interfaces';
 
-export async function handleWebhookEvent(webhookData: {
-  id: string;
-  createdAt: string;
-  event: WebhookEvent;
-  spaceId: string;
-}) {
-  switch (webhookData.event.scope) {
-    case WebhookEventNames.DocumentMentionCreated: {
-      const mentionedUserId = webhookData.event.mention.value;
-      const mentionAuthorId = webhookData.event.user.id;
-
-      if (mentionedUserId !== mentionAuthorId) {
-        await createDocumentNotification({
-          type: 'mention.created',
-          createdBy: mentionAuthorId,
-          mentionId: webhookData.event.mention.id,
-          pageId: webhookData.event.document.id,
-          spaceId: webhookData.spaceId,
-          userId: mentionedUserId
-        });
-      }
-
-      break;
-    }
-
-    case WebhookEventNames.DocumentInlineCommentCreated: {
-      return createInlineCommentNotification(webhookData.event);
-    }
-
-    case WebhookEventNames.CardInlineCommentCreated: {
-      return createInlineCommentNotification(webhookData.event);
-    }
-
-    case WebhookEventNames.CardMentionCreated: {
-      const mentionedUserId = webhookData.event.mention.value;
-      const mentionAuthorId = webhookData.event.user.id;
-
-      await createCardNotification({
-        type: 'mention.created',
-        createdBy: mentionAuthorId,
-        mentionId: webhookData.event.mention.id,
-        cardId: webhookData.event.card.id,
-        spaceId: webhookData.spaceId,
-        userId: mentionedUserId
-      });
-
-      break;
-    }
-
-    case WebhookEventNames.CardBlockCommentCreated: {
-      const spaceId = webhookData.spaceId;
-      const commentAuthorId = webhookData.event.blockComment.author.id;
-      const cardId = webhookData.event.card.id;
-      const blockCommentId = webhookData.event.blockComment.id;
-      const cardAuthorId = webhookData.event.card.author.id;
-
-      const blockComment = await prisma.block.findFirstOrThrow({
-        where: {
-          id: blockCommentId
-        },
-        select: {
-          parentId: true,
-          fields: true
-        }
-      });
-
-      const blockCommentContent: PageContent = (blockComment.fields as any).content;
-
-      const previousComment = await prisma.block.findFirst({
-        where: {
-          parentId: blockComment.parentId,
-          type: 'comment'
-        },
-        orderBy: {
-          createdAt: 'desc'
-        },
-        skip: 1,
-        take: 1,
-        select: {
-          id: true,
-          createdBy: true
-        }
-      });
-
-      const previousCommentId = previousComment?.id;
-      const previousCommentAuthorId = previousComment?.createdBy;
-
-      if (cardAuthorId !== commentAuthorId) {
-        await createCardNotification({
-          type: 'block_comment.created',
-          createdBy: commentAuthorId,
-          cardId,
-          spaceId,
-          userId: cardAuthorId,
-          blockCommentId
-        });
-      }
-
-      if (
-        previousCommentId &&
-        previousCommentAuthorId &&
-        previousCommentAuthorId !== commentAuthorId &&
-        cardAuthorId !== commentAuthorId
-      ) {
-        await createCardNotification({
-          type: 'block_comment.replied',
-          createdBy: commentAuthorId,
-          cardId,
-          spaceId,
-          userId: previousCommentAuthorId,
-          blockCommentId
-        });
-      }
-
-      const extractedMentions = extractMentions(blockCommentContent);
-
-      for (const extractedMention of extractedMentions) {
-        const mentionedUserId = extractedMention.value;
-        await createCardNotification({
-          type: 'block_comment.mention.created',
-          createdBy: commentAuthorId,
-          cardId,
-          mentionId: extractedMention.id,
-          spaceId,
-          userId: mentionedUserId,
-          blockCommentId
-        });
-      }
-
-      break;
-    }
-
-    case WebhookEventNames.CardPersonPropertyAssigned: {
-      const spaceId = webhookData.spaceId;
-      const assignedUserId = webhookData.event.assignedUser.id;
-      const cardId = webhookData.event.card.id;
-
-      await createCardNotification({
-        type: 'person_assigned',
-        personPropertyId: webhookData.event.personPropertyId,
-        cardId,
-        spaceId,
-        userId: assignedUserId,
-        createdBy: webhookData.event.user.id
-      });
-      break;
-    }
-
-    default:
-      break;
-  }
-}
 /**
  * SQS worker, message are executed one by one
  */
@@ -180,25 +27,27 @@ export const webhookWorker = async (event: SQSEvent): Promise<SQSBatchResponse> 
         // Gets message information
         const { webhookURL, signingSecret, ...webhookData } = JSON.parse(body) as WebhookPayload;
 
-        const webhookEventId = webhookData.id;
+        const jsonString = JSON.stringify(webhookData);
+
+        const webhookMessageHash = crypto.createHash('sha256').update(jsonString).digest('hex');
 
         const webhookMessage = await prisma.webhookMessage.findUnique({
           where: {
-            id: webhookEventId,
+            id: webhookMessageHash,
             processed: true
           }
         });
 
         if (webhookMessage) {
-          log.debug('Webhook message already processed', { webhookEventId, spaceId: webhookData.spaceId });
+          log.debug('Webhook message already processed', { webhookData });
           return;
         }
 
-        await prisma.webhookMessage.update({
-          where: {
-            id: webhookEventId
-          },
+        await createNotificationsFromEvent(webhookData);
+
+        await prisma.webhookMessage.create({
           data: {
+            id: webhookMessageHash,
             processed: true
           }
         });
@@ -208,7 +57,7 @@ export const webhookWorker = async (event: SQSEvent): Promise<SQSBatchResponse> 
         if (!isWhitelistedEvent) {
           log.debug('Webhook event not whitelisted', {
             scope: webhookData.event.scope,
-            webhookEventId,
+            webhookData,
             spaceId: webhookData.spaceId
           });
           return;
